src/main/java/cothe/webservice/service/PostsService.java
package cothe.webservice.service;

import cothe.webservice.domain.posts.PostsRepository;
import cothe.webservice.web.PostsSaveRequestDto;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

/**
 * @author Jeongjin Kim
 * @since 2018. 2. 14.
 */
@AllArgsConstructor
@Service
public class PostsService {
    private PostsRepository postsRepository;

    @Transactional
    public long save(PostsSaveRequestDto dto){
        return postsRepository.save(dto.toEntity()).getId();
    }
}
